Move write_packageurls_to_file and write_data_to_file to minecode_pipeline.pipes #660
